feat: implement ext2 sparse file support — read holes as zero-filled blocks (Item 148)
Ext2 sparse files (files with unallocated holes) are now handled
correctly during read operations:

ext2_get_block_num() helper:
- Refactored block number resolution into a standalone helper
  that returns the physical block number (0 = hole) for any
  logical block index
- Supports direct blocks (0-11), singly indirect blocks, and
  correctly identifies holes in both direct and indirect maps
- Returns -1 for doubly/triply indirect (not yet supported)

ext2_read_inode_block():
- Uses ext2_get_block_num() for resolution
- When encountering a hole (phys_block == 0), zero-fills the
  output buffer via memset() instead of returning -1
- Hole blocks are silently filled with zeros per POSIX semantics

Also fixed pre-existing -Werror warnings across the codebase:
- syscall.c: cast pointers to uint64_t in syscall_user_read_ok()
- yama.h: add yama_sysctl_register() declaration
- cpuidle.c: fix nested /* comment
- process.c: add KERNEL_INTERNAL define for prng_rand64()
- mutex.c: remove unused owner_on_cpu variable

+++ src/kernel/syscall.c
@@ -8575,7 +8575,7 @@ static uint64_t sys_name_to_handle_at(uint64_t dirfd, uint64_t pathname,
     /* Read the current handle_bytes from user space */
     unsigned int handle_bytes;
     if (syscall_is_user_process()) {
-        if (!syscall_user_read_ok(&handle->handle_bytes, sizeof(handle->handle_bytes)))
+        if (!syscall_user_read_ok((uint64_t)(uintptr_t)&handle->handle_bytes, sizeof(handle->handle_bytes)))
             return (uint64_t)(int64_t)-EFAULT;
     }
     handle_bytes = handle->handle_bytes;
@@ -8649,7 +8649,7 @@ static uint64_t sys_open_by_handle_at(uint64_t mount_fd, uint64_t handle_addr,
     /* Read handle_bytes from userspace */
     unsigned int hb;
     if (syscall_is_user_process()) {
-        if (!syscall_user_read_ok(&handle->handle_bytes, sizeof(handle->handle_bytes)))
+        if (!syscall_user_read_ok((uint64_t)(uintptr_t)&handle->handle_bytes, sizeof(handle->handle_bytes)))
             return (uint64_t)(int64_t)-EFAULT;
     }
     hb = handle->handle_bytes;
